requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Data validation and models
pydantic>=2.0.0

# YAML configuration
pyyaml>=6.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and audit
structlog>=23.0.0

# Testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (development)
mypy>=1.5.0

# Code formatting (development)
black>=23.0.0
flake8>=6.0.0

# OCR & Image Processing
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Multilingual (for Phase 2)
indic-transliteration>=2.3.0
langdetect>=1.0.9

# ASR (for Phase 2)
SpeechRecognition>=3.10.0
pydub>=0.25.1

# AI/LLM (for Phase 3)
openai>=1.0.0
anthropic>=0.18.0

# Web API (for Phase 4)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
